requirements.txt
## Django dependencies

# django with argon2 support
django[argon2]==4.1.3
# django load .env file
django-environ==0.9.0
# Database clients libraries
psycopg2-binary==2.9.4
mysqlclient==2.1.1

# Django rest framework libraries
djangorestframework==3.14.0
# filter rest parameters
django-filter==22.1

# dj-rest-auth
djangorestframework-simplejwt==5.2.2
django-vite==2.0.2
django-cors-headers==3.13.0

# For cryptography
cryptography==38.0.3

# WSGI compliant production server
gunicorn==20.1.0
Pillow==9.2.0
